requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
reportlab==4.0.7
pandas==2.1.4
python-dotenv==1.0.0
openpyxl==3.1.2
PyPDF2==3.0.1
transformers==4.36.0
torch==2.1.0
